+++ tests/FileService.Tests/FileSystemTests.cs
@@ -67,7 +67,7 @@ public class FileSystemTests
         const string searchPath = "searchPath";
         const string ignorePath = "ignorePath";
 
-        // In FileSystem, `GetFilesAsync` returns files in alphabetical order, so this naming makes the assertions simpler.
+        // In FileSystem, `ListFilesAsync` returns files in alphabetical order, so this naming makes the assertions simpler.
         // (In InMemory, files are returned in the order added, so naming doesn't matter.)
         var files = new List<(string path, string fileName)>
         {
@@ -82,7 +82,7 @@ public class FileSystemTests
 
         // Act
         // -- Only find files starting with `searchPath`.
-        var results = _fileService.GetFilesAsync(searchPath);
+        var results = _fileService.ListFilesAsync(searchPath);
 
         // Assert
         var i = 0;
@@ -102,7 +102,7 @@ public class FileSystemTests
         const string searchPath = "searchPath";
         const string nestedPath = "nestedPath";
 
-        // In FileSystem, `GetFilesAsync` returns files in alphabetical order, so this naming makes the assertions simpler.
+        // In FileSystem, `ListFilesAsync` returns files in alphabetical order, so this naming makes the assertions simpler.
         // (In InMemory, files are returned in the order added, so naming doesn't matter.)
         var files = new List<(string path, string fileName)>
         {
@@ -117,7 +117,7 @@ public class FileSystemTests
 
         // Act
         // -- Only find files starting with `searchPath`.
-        var results = _fileService.GetFilesAsync(searchPath);
+        var results = _fileService.ListFilesAsync(searchPath);
 
         // Assert
         var i = 0;
@@ -134,7 +134,7 @@ public class FileSystemTests
     public async Task GetFiles_WhenFileDoesNotExist_ReturnsEmptyList()
     {
         // Act
-        var results = _fileService.GetFilesAsync();
+        var results = _fileService.ListFilesAsync();
 
         // Assert
         await foreach (var _ in results) Assert.Fail("`results` should be empty.");
@@ -145,7 +145,7 @@ public class FileSystemTests
     public async Task GetFiles_WhenPathDoesNotExist_ReturnsEmptyList()
     {
         // Act
-        var results = _fileService.GetFilesAsync("nope");
+        var results = _fileService.ListFilesAsync("nope");
 
         // Assert
         await foreach (var _ in results) Assert.Fail("`results` should be empty.");